apps/studio-api/pyproject.toml
[project]
name = "wisdom-studio-api"
version = "0.6.0"
description = "FastAPI backend for Wisdom Studio — transport-only over the Wisdom Layer SDK."
readme = "../../README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Rhatigan AGI" }]
keywords = ["wisdom-layer", "agent", "ai", "studio", "reference"]
dependencies = [
    "wisdom-layer>=1.1.0,<2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "websockets>=13.0",
    "pyyaml>=6.0",
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",
    "bleach>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.50.0"]
gemini = ["google-genai>=0.3.0"]
ollama = ["ollama>=0.4.0"]
litellm = ["litellm>=1.50.0"]
all-adapters = [
    "openai>=1.50.0",
    "google-genai>=0.3.0",
    "ollama>=0.4.0",
    "litellm>=1.50.0",
]
postgres = ["asyncpg>=0.29.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["studio_api"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM"]
ignore = ["E501"]
# TC* rules push imports into TYPE_CHECKING blocks; FastAPI's Depends() and
# Pydantic field resolution need runtime imports, so we keep TC off.

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
